requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2023.5.7
ephem==4.1.4
PySocks==1.7.1
python-telegram-bot==13.14
pytz==2023.3
six==1.16.0
tornado==6.1
tzdata==2023.3
tzlocal==5.0.1
emoji==2.4.0